+++ cloud/objstore/objcli/errors.go
@@ -2,7 +2,6 @@ package objcli
 
 import (
 	"errors"
-	"fmt"
 )
 
 var (
@@ -14,11 +13,3 @@ var (
 	// require upload ids.
 	ErrExpectedNoUploadID = errors.New("received an unexpected upload id, cloud provider doesn't required upload ids")
 )
-
-type ErrVersionUnexpectedType struct {
-	TypeName string
-}
-
-func (e ErrVersionUnexpectedType) Error() string {
-	return fmt.Sprintf("the version field is of an unexpected type, expected %s", e.TypeName)
-}