+++ cloud/objstore/objcli/objazure/api.go
@@ -29,9 +29,9 @@ func (c *serviceClient) NewContainerClient(containerName string) containerAPI {
 }
 
 type containerAPI interface {
-	NewBlobClient(blobName string) blobAPI
-	NewBlockBlobClient(blobName string) blockBlobAPI
-	NewBlockBlobVersionClient(blobName, versionID string) (blockBlobAPI, error)
+	NewBlobClient(name string) blobAPI
+	NewBlockBlobClient(name string) blockBlobAPI
+	NewBlockBlobVersionClient(name, version string) (blockBlobAPI, error)
 	NewListBlobsFlatPager(o *container.ListBlobsFlatOptions) flatBlobsPager
 	NewListBlobsHierarchyPager(delimiter string, o *container.ListBlobsHierarchyOptions) hierarchyBlobsPager
 }
@@ -40,18 +40,16 @@ type containerClient struct {
 	client *container.Client
 }
 
-// NewBlockBlobVersionClient creates a blob block client. It creates a version specific client if the versionID is
-// provided, if not it creates a normal client.
-func (c containerClient) NewBlockBlobVersionClient(blobName, versionID string) (blockBlobAPI, error) {
-	return c.client.NewBlockBlobClient(blobName).WithVersionID(versionID)
+func (c containerClient) NewBlockBlobClient(name string) blockBlobAPI {
+	return c.client.NewBlockBlobClient(name)
 }
 
-func (c containerClient) NewBlockBlobClient(blobName string) blockBlobAPI {
-	return c.client.NewBlockBlobClient(blobName)
+func (c containerClient) NewBlockBlobVersionClient(name, version string) (blockBlobAPI, error) {
+	return c.client.NewBlockBlobClient(name).WithVersionID(version)
 }
 
-func (c containerClient) NewBlobClient(blobName string) blobAPI {
-	return c.client.NewBlobClient(blobName)
+func (c containerClient) NewBlobClient(name string) blobAPI {
+	return c.client.NewBlobClient(name)
 }
 
 func (c containerClient) NewListBlobsFlatPager(o *container.ListBlobsFlatOptions) flatBlobsPager {