+++ cloud/objstore/objutil/compress_upload_test.go
@@ -4,11 +4,11 @@ import (
 	"archive/zip"
 	"bytes"
 	"context"
+	"crypto/rand"
 	"crypto/sha256"
 	"io"
 	"log/slog"
 	"math"
-	"math/rand"
 	"regexp"
 	"testing"